(svn r17464) -Codechange: use C++0x mode when compiling with gcc >= 4.3 or icc >= 11.0

+++ src/table/settings.h
@@ -83,7 +83,7 @@ static bool UpdateClientConfigValues(int32 p1);
  */
 
 #define NSD_GENERAL(name, def, cmd, guiflags, min, max, interval, many, str, proc, load)\
-	{name, (const void*)(def), {cmd}, {guiflags}, min, max, interval, many, str, proc, load}
+	{name, (const void*)(def), {(byte)cmd}, {(byte)guiflags}, min, max, interval, many, str, proc, load}
 
 /* Macros for various objects to go in the configuration file.
  * This section is for global variables */
